+++ packages/react/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ignite-ui-vitinho/react
 
+## 1.1.0
+
+### Minor Changes
+
+- Add Tooltip and Toast components
+
 ## 1.0.1
 
 ### Patch Changes