+++ packages/tokens/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ignite-ui-vitinho/tokens
 
+## 1.2.0
+
+### Minor Changes
+
+- Add Tooltip and Toast components
+
 ## 1.1.0
 
 ### Minor Changes